Remove sourcecompile
The "sourcecompile" module was used to emulate the feature of
python 2.6's compile() builtin function allowing it to operate
on AST objects.  It was originally my intent for the code to be
executed as follows:

   * parse
   * transform (print statement into print function)
   * verify (using cappython)
   * compile
   * execute

But on GAE the compile() step is not allowed on AST objects even
if you pull in an implementation of compile() from python 2.6 which
is compatible with a standard python 2.5 interpreter because it
is not allowed to create code objects in restricted execution node.

To work around this, I changed the execution method to something like
the following:

   * parse
   * transform
   * verify
   * generate altered sources
   * compile
   * execute

This commit changes it to perform the generation of altered sources
at an earlier stage, something like:

   * parse
   * transform
   * generate altered sources
   * parse altered sources
   * verify
   * compile altered sources
   * execute

This would seem to be slightly less efficient but it has two advantages.
First, it performs the generation of source code before the verification
step so it is less likely that a bug in the code generation component
will compromise the sandbox.  Second, cappython's safeeval module is
now synchronised with upstream except for the custom import lines.