Added --java option (not implemented yet)
I am going to try to deploy the same application using the Java version of GAE.
As well as being interested in whether this is possible there is a particular
java library, caja, that I may be interested in using.  I am also quite
excited about Jython reaching version 2.5 and would like to put it through
its paces.